requirements.txt
functions-framework==3.*
google-cloud-storage
pandas
fsspec
gcsfs